requirements.txt
peewee==3.14.8
python-dotenv==0.19.2
pyTelegramBotAPI==4.3.1
Flask==2.0.2
loguru==0.6.0
python-telegram-bot-calendar==1.0.5
requests==2.27.1